// src/api/generated/models/companyProjectResource.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * brew
 * OpenAPI spec version: 0.0.1
 */
import type { CompanyProjectResourceTitlePt } from './companyProjectResourceTitlePt';
import type { CompanyProjectResourceTitleEn } from './companyProjectResourceTitleEn';
import type { CompanyProjectResourceDescriptionPt } from './companyProjectResourceDescriptionPt';
import type { CompanyProjectResourceDescriptionEn } from './companyProjectResourceDescriptionEn';
import type { CompanyProjectResourceTranslationStatusLabel } from './companyProjectResourceTranslationStatusLabel';
import type { CompanyProjectResourceProdUrl } from './companyProjectResourceProdUrl';
import type { CompanyProjectResourceGithubUrl } from './companyProjectResourceGithubUrl';
import type { CompanyProfileResource } from './companyProfileResource';
import type { LanguageResource } from './languageResource';
import type { CompanyProjectResourceCreatedAt } from './companyProjectResourceCreatedAt';
import type { CompanyProjectResourceUpdatedAt } from './companyProjectResourceUpdatedAt';

export interface CompanyProjectResource {
  id: string;
  title: string;
  title_pt: CompanyProjectResourceTitlePt;
  title_en: CompanyProjectResourceTitleEn;
  description: string;
  description_pt: CompanyProjectResourceDescriptionPt;
  description_en: CompanyProjectResourceDescriptionEn;
  translation_status: string;
  translation_status_label: CompanyProjectResourceTranslationStatusLabel;
  prod_url: CompanyProjectResourceProdUrl;
  github_url: CompanyProjectResourceGithubUrl;
  company_profile_id: string;
  company_profile?: CompanyProfileResource;
  languages?: LanguageResource[];
  created_at: CompanyProjectResourceCreatedAt;
  updated_at: CompanyProjectResourceUpdatedAt;
}